+++ src/com/intellij/idea/plugin/hybris/settings/HybrisProjectSettings.kt
@@ -50,4 +50,5 @@ class HybrisProjectSettings : BaseState() {
     var availableExtensions by property(TreeMap<String, ExtensionDescriptor> { a, b -> a.compareTo(b, true) }) { it.isEmpty() }
     var excludedFromScanning by stringSet()
     var remoteConnectionSettingsList by list<HybrisRemoteConnectionSettings>()
+    var useFakeOutputPathForCustomExtensions by property(false)
 }